move download merged project feed into MonitorableJob
